+++ engine/core/window.cpp
@@ -48,7 +48,7 @@ void ce::Window::swapBuffers() {
 glm::vec2 ce::Window::getWindowSize() {
 	int w, h;
 	SDL_GetWindowSize(m_window, &w, &h);
-	return glm::vec2((float)w, (float)h);
+	return glm::vec2(w, h);
 }
 float ce::Window::getAspectRatio() {
 	glm::vec2 size = getWindowSize();
